Add NVMe LBA format check
Checks for any mismatches in block size between NVMe devices, and
individually checks metadata size (if not 0), block size (if larger than
4K), relative performance (if not 0).

If block size is different but 4K or less, only a warning is triggered.
If any of the other checks fail, a failure is triggered.

Note that we don't check if the same specific LBA format is common (i.e.
the actual LBA format number or ID) as this will likely be different for
different NVMe models, and could even be different between revisions.
While the model would ideally be the same, in reality they may not be.