Create basic scaffold for comparing output of a single command across multiple servers
Often we want to compare the results of one command across multiple backends because
a single server being different is likely to indicate a misconfiguration.
This code takes the standard out of commands that are defined with
SCRIPT_TYPE="parallel-compare-backends"
and compares them. A single differing output is enough to be considered a failure.

Scripts of this type should therefore take steps to ensure their output is
in a predictable consistent order: liberal use of $(sort) is probably helpful.