Move to version table-based detection of errors
For a lot of these issues the only possible test is "are you running
a susceptible version?" and it's impossible to further check whether
the actual issue will be encountered without external knowledge.
This table makes it easy to do a simple check on versions and report
a matching bug report and advisory note.