Added a TEST-only version of a script to evaluate source-based routing
In WEKAPP-360289 we discovered a cluster that had two IPs on the same
subnet, but the hosts' device-specific route tables were not set up to
route to the local subnet.
This meant that the device-specific route tables were basically ignored
for the local subnet, and it was as if no SBR was defined.

This is a draft version of a script that simply attempts to count routes
to the relevant destination in the hope that this will catch the above
situation.
Configuring SBR is a flexible process, and it can be a difficult task
to test for, so this definitely won't catch every misconfiguration.