Added 3 more scripts to check for problematic protocol configurations
Protocols which use floating IP addresses, like NFS-W, SMB-W and S3
rely on certain things in the Linux kernel.
They need to be able to determine routes and devices correctly,
and respond to requests to those floating IPs.

These scripts are very basic attempts to check that the "normal"
conditions exist.