Replace calls to jq with calls to python3 + json module
At the moment we can only call shell scripts in wekachecker. This
means that the natural solution to parse JSON is jq, but we can't
rely on jq being available on default installations of older
distributions (such as RHEL7).

As Python3 is more likely to be more widely installed (and these
changes only require the sys, json, and collection modules, which
are expected to be stable) it seems the only acceptable change
is to remove the dependency on jq and move to Python3.

This has led to some less-readable code (in particular in the
script 460_ip_source-based_routing.sh) so improvements are welcome.