Added a very simple test to see if OBS tiers are in "Scarce" mode
In WEKAPP-376458 there was a problem with NFS services getting
stuck. The issue went through Protocols -> Driver -> Filesystem
team before it was found that an OBS tier was in scarce mode,
limiting backpressure's ability to free up NVMe space.

While this is not a diagnostic certainty, this simple test might
direct investigation more accurately.