Add a check to verify that /opt/weka is configured to mount at boot time
in SFDC 13589 we encountered a situation in which the customer had
configured an extra partition for /opt/weka, and defined it in /etc/fstab
but had not actually mounted it.
All changes were made to /opt/weka on the rootFS.

When the system rebooted the weka containers were missing.

This script checks for the definition of an /opt/weka mountpoint in
both fstab and systemd, and whether it's currently mounted (and the
inverse).

This is not foolproof because you could still achieve the same
confusion by e.g. doing similar with /opt rather than /opt/weka