Allow hostname to match on shortname, bail out of testing if no frontend0 container
Sometimes a Weka cluster will have hostnames defined as "shortname", rather than
the full FQDN output of $(hostname). Therefore we should not use it as a strict
filter, but instead grep -w on the short hostname

Also, if we couldn't find a matching frontend0 container, there's little point
checking the local host's routing tables, because it should not be involved in
NFS, so bail out early.