Limit warning to only those cases where the routing destinations match
The previous version warned simply if we had metrics set, but that's not
precise enough; metrics only really matter if the routes have the same
destination. Even then there has to be more than one metric-having
route to the same destination for us to actually have a problem, therefore:

First find any routes that specify a route metric, then count them based
on the same routing destination. A single route to a destination with
a metric is fine (because the metric is meaningless there), but routes
which specify the same destination *and* a metric is not likely to be useful.
(There are circumstances where it makes sense, such as a backup/failover
 route that's not intended to be preferred)