First pass at framework to diagnose performance problems based on specific criteria
If we can classify problem conditions by specific criteria in stats
and then check to see if those values apply, maybe we can start
diagnosing or at least pointing to hardware problems.

The basic theory being (e.g. for a faulty NVME):
    If one disk has read latency > 5ms
    And cluster average read latency > 100usec
    And ...
    And ...

Then maybe we can build a set of test parameters for each one of
those outcomes.
This is not going to be a small undertaking, and unfortunately
we're stuck with bash. However, if the outline test conditions
and theory can be proven useful, this is a step towards a useful
enhancement.

Ideally the logic/tests that come out of this would make it into
the product somehow.