Check all mdadm devices report as OK
The WSA will auto-create mdadm RAID1 devices if there are two small
hardware drives, and use those as boot devices. In SFDC 00018153
some drives got out of sync due to a firmware update, and nothing
reports on that. The symptom was that changes in /boot where not
being reflected in the boot loader because the device GRUB was
loading from wasn't being updated.