Add a very basic capacity check in the case of 0 hot-spare clusters
Clusters with 0 hot-spare can potentially be configured to allocate
and use ~100% of SSD capacity. In the case of loss of a failure
domain they'll lose a proportion of FS space (known as shrinkage).
The proportion lost is dependent on the number of data disks, and
if "too much" is lost then writes can fail with ENOSPACE.

Essentially, we should warn customers without hot spares configured