Basic checker to compare current NIC link speed with maximum
In https://wekaio.atlassian.net/browse/WEKAPP-482528 we saw that
the link speed was lower than expected, but there were no warnings.
We should check that.

The only plausible way I can find of doing this is by parsing the
text-based output of ethtool, until jq and "ethtool --json" get
everywhere, or the kernel interface to ethtool-netlink is exposed
in /sys, I can't see any other way of doing it. :(