Move from arp_ignore=0 to arp_ignore=1
In August 2025 we saw that -- for reasons unknown -- docs (and
thus TA Tool) recommended arp_ignore=0. This is unwise as it
allows the Linux kernel to respond to ARP requests for NIC A via
NIC B, which has a habit of getting switches to learn the wrong
MAC Address for their CAM tables.

https://wekaio.slack.com/archives/C05QSG8HCF9/p1755274686557269
https://wekaio.slack.com/archives/C03E96QFM33/p1755509369672719